[V1][Metrics] Replace prefix cache hit rate with queries/hits counters
With a hit rate metric, we need to avoid the situation where if the hit
rate drops due to a sudden increase in misses, the change gets diluted
by the large accumulated totals.

With Prometheus, it is better to log monotonic counters and use
Prometheus queries to calculate rates according to whatever time
interval makes sense to the operator, also allowing them to easily
adjust the time interval in their PromQL queries.

With the logging logger, we can continue to aggregate across an
interval of (roughly) 1000 requests.

Signed-off-by: Mark McLoughlin <markmc@redhat.com>