[V1] Refactor parallel sampling support
The initial implementation went to great efforts to add parallel
sampling as a wrapper at the highest layer of abstraction possible.
This resulted in a lot of tricky code to post-process RequestOutputs
to aggregate them where necessary.

Instead, it probably makes sense to implement parallel sampling at
the layer that actually creates RequestOutput objects - i.e. in
OutputProcessor.

To do this, we simply need to allow for fanning out child requests
in LLMEngine.add_request(), passing details of the fan-out to
OutputProcessor.

This adds some overhead to the n=1 case (see SingularSamplingRequest)
in return for significantly less overhead and complication in the
parallel sampling case.

Signed-off-by: Mark McLoughlin <markmc@redhat.com>